requirements.txt
streamlit==1.31.1
diffusers==0.21.4
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
pillow==10.1.0
huggingface-hub==0.16.4 